High availability in a multi-node Siddhi deployment
This commit implements the high availability for a multi-node Siddhi deployment
using Kafka transport as the communication medium between nodes.

eg:
kafka -> Siddhi1 -> kafka -> Siddhi2 -> output

Feature is tested using a test case, which creates above shown scenario
and executes a siddhi file in a distributed manner (1 query is deployed
in Siddhi1 and the other query is deployed in Siddhi2).